feat(ui): let eraser tool handle its events
